--- app/src/main/java/com/dicoding/tourismapp/core/data/TourismRepository.kt
@@ -1,58 +1,0 @@
-package com.dicoding.tourismapp.core.data
-
-import com.dicoding.tourismapp.core.data.source.local.LocalDataSource
-import com.dicoding.tourismapp.core.data.source.remote.RemoteDataSource
-import com.dicoding.tourismapp.core.data.source.remote.network.ApiResponse
-import com.dicoding.tourismapp.core.data.source.remote.response.TourismResponse
-import com.dicoding.tourismapp.core.domain.model.Tourism
-import com.dicoding.tourismapp.core.domain.repository.ITourismRepository
-import com.dicoding.tourismapp.core.utils.AppExecutors
-import com.dicoding.tourismapp.core.utils.DataMapper
-import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
-import io.reactivex.rxjava3.core.Flowable
-import io.reactivex.rxjava3.schedulers.Schedulers
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class TourismRepository @Inject constructor(
-    private val remoteDataSource: RemoteDataSource,
-    private val localDataSource: LocalDataSource,
-    private val appExecutors: AppExecutors
-) : ITourismRepository {
-    override fun getAllTourism(): Flowable<Resource<List<Tourism>>> =
-        object : NetworkBoundResource<List<Tourism>, List<TourismResponse>>(appExecutors) {
-            override fun loadFromDB(): Flowable<List<Tourism>> {
-                return localDataSource.getAllTourism().map {
-                    DataMapper.mapEntitiesToDomain(it)
-                }
-            }
-
-            override fun shouldFetch(data: List<Tourism>?): Boolean {
-                return data.isNullOrEmpty()
-            }
-
-            override fun createCall(): Flowable<ApiResponse<List<TourismResponse>>> =
-                remoteDataSource.getAllTourism()
-
-            override fun saveCallResult(data: List<TourismResponse>) {
-                val tourismList = DataMapper.mapResponsesToEntities(data)
-                localDataSource.insertTourism(tourismList)
-                    .subscribeOn(Schedulers.io())
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .subscribe()
-            }
-        }.asFlowable()
-
-    override fun getFavoriteTourism(): Flowable<List<Tourism>> {
-        return localDataSource.getFavoriteTourism().map {
-            DataMapper.mapEntitiesToDomain(it)
-        }
-    }
-
-    override fun setFavoriteTourism(tourism: Tourism, state: Boolean) {
-        val tourismEntity = DataMapper.mapDomainToEntity(tourism)
-        appExecutors.diskIO().execute { localDataSource.setFavoriteTourism(tourismEntity, state) }
-    }
-}
-
