--- app/src/main/java/com/dicoding/tourismapp/core/data/source/remote/network/ApiResponse.kt
@@ -1,7 +1,0 @@
-package com.dicoding.tourismapp.core.data.source.remote.network
-
-sealed class ApiResponse<out R> {
-    data class Success<out T>(val data: T) : ApiResponse<T>()
-    data class Error(val errorMessage: String) : ApiResponse<Nothing>()
-    data object Empty : ApiResponse<Nothing>()
-}
