--- app/src/main/java/com/dicoding/tourismapp/core/data/source/remote/response/ListTourismResponse.kt
@@ -1,14 +1,0 @@
-package com.dicoding.tourismapp.core.data.source.remote.response
-
-import com.google.gson.annotations.SerializedName
-
-data class ListTourismResponse(
-    @field:SerializedName("error")
-    val error: Boolean,
-
-    @field:SerializedName("message")
-    val message: String,
-
-    @field:SerializedName("places")
-    val places: List<TourismResponse>
-)
