--- app/src/main/java/com/dicoding/tourismapp/core/di/AppModule.kt
@@ -1,17 +1,0 @@
-package com.dicoding.tourismapp.core.di
-
-import com.dicoding.tourismapp.core.domain.interactor.TourismInteractor
-import com.dicoding.tourismapp.core.domain.usecase.TourismUseCase
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
-import dagger.hilt.android.scopes.ViewModelScoped
-
-@Module
-@InstallIn(ViewModelComponent::class)
-abstract class AppModule {
-    @Binds
-    @ViewModelScoped
-    abstract fun provideTourismUseCase(tourismInteractor: TourismInteractor): TourismUseCase
-}
