--- app/src/main/java/com/dicoding/tourismapp/core/di/RepositoryModule.kt
@@ -1,15 +1,0 @@
-package com.dicoding.tourismapp.core.di
-
-import com.dicoding.tourismapp.core.data.TourismRepository
-import com.dicoding.tourismapp.core.domain.repository.ITourismRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module(includes = [NetworkModule::class, DatabaseModule::class])
-@InstallIn(SingletonComponent::class)
-abstract class RepositoryModule {
-    @Binds
-    abstract fun provideRepository(tourismRepository: TourismRepository): ITourismRepository
-}
