--- app/src/main/java/com/dicoding/tourismapp/core/domain/interactor/TourismInteractor.kt
@@ -1,23 +1,0 @@
-package com.dicoding.tourismapp.core.domain.interactor
-
-import com.dicoding.tourismapp.core.data.Resource
-import com.dicoding.tourismapp.core.domain.model.Tourism
-import com.dicoding.tourismapp.core.domain.repository.ITourismRepository
-import com.dicoding.tourismapp.core.domain.usecase.TourismUseCase
-import io.reactivex.rxjava3.core.Flowable
-import javax.inject.Inject
-
-class TourismInteractor @Inject constructor(private val tourismRepository: ITourismRepository) : TourismUseCase {
-    override fun getAllTourism(): Flowable<Resource<List<Tourism>>> {
-        return tourismRepository.getAllTourism()
-    }
-
-    override fun getFavoriteTourism(): Flowable<List<Tourism>> {
-        return tourismRepository.getFavoriteTourism()
-    }
-
-    override fun setFavoriteTourism(tourism: Tourism, state: Boolean) {
-        return tourismRepository.setFavoriteTourism(tourism, state)
-    }
-
-}
