--- app/src/main/java/com/dicoding/tourismapp/core/utils/AppExecutors.kt
@@ -1,41 +1,0 @@
-package com.dicoding.tourismapp.core.utils
-
-import android.os.Handler
-import android.os.Looper
-import androidx.annotation.VisibleForTesting
-
-import java.util.concurrent.Executor
-import java.util.concurrent.Executors
-import javax.inject.Inject
-
-class AppExecutors @VisibleForTesting constructor(
-    private val diskIO: Executor,
-    private val networkIO: Executor,
-    private val mainThread: Executor
-) {
-
-    companion object {
-        private const val THREAD_COUNT = 3
-    }
-
-    @Inject
-    constructor() : this(
-        Executors.newSingleThreadExecutor(),
-        Executors.newFixedThreadPool(THREAD_COUNT),
-        MainThreadExecutor()
-    )
-
-    fun diskIO(): Executor = diskIO
-
-    fun networkIO(): Executor = networkIO
-
-    fun mainThread(): Executor = mainThread
-
-    private class MainThreadExecutor : Executor {
-        private val mainThreadHandler = Handler(Looper.getMainLooper())
-
-        override fun execute(command: Runnable) {
-            mainThreadHandler.post(command)
-        }
-    }
-}
