+++ core/src/main/java/app/culturedev/core/data/Resource.kt
@@ -1,0 +1,8 @@
+package app.culturedev.core.data
+
+sealed class Resource<T>(val data: T? = null, val message: String? = null) {
+    class Success<T>(data: T) : Resource<T>(data)
+    class Loading<T>(data: T? = null) : Resource<T>(data)
+    class Error<T>(message: String, data: T? = null) :
+        Resource<T>(data, message)
+}
