+++ core/src/main/java/app/culturedev/core/data/source/local/room/TourismDao.kt
@@ -1,0 +1,26 @@
+package app.culturedev.core.data.source.local.room
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import androidx.room.Update
+import app.culturedev.core.data.source.local.entity.TourismEntity
+import io.reactivex.rxjava3.core.Completable
+import io.reactivex.rxjava3.core.Flowable
+
+@Dao
+interface TourismDao {
+
+    @Query("SELECT * FROM tourism")
+    fun getAllTourism(): Flowable<List<TourismEntity>>
+
+    @Query("SELECT * FROM tourism where isFavorite = 1")
+    fun getFavoriteTourism(): Flowable<List<TourismEntity>>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun insertTourism(tourism: List<TourismEntity>): Completable
+
+    @Update
+    fun updateFavoriteTourism(tourism: TourismEntity)
+}
