+++ core/src/main/java/app/culturedev/core/data/source/remote/RemoteDataSource.kt
@@ -1,0 +1,36 @@
+package app.culturedev.core.data.source.remote
+
+import android.annotation.SuppressLint
+import app.culturedev.core.data.source.remote.network.ApiResponse
+import app.culturedev.core.data.source.remote.network.ApiService
+import app.culturedev.core.data.source.remote.response.TourismResponse
+import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
+import io.reactivex.rxjava3.core.BackpressureStrategy
+import io.reactivex.rxjava3.core.Flowable
+import io.reactivex.rxjava3.schedulers.Schedulers
+import io.reactivex.rxjava3.subjects.PublishSubject
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class RemoteDataSource @Inject constructor(private val apiService: ApiService) {
+    @SuppressLint("CheckResult")
+    fun getAllTourism(): Flowable<ApiResponse<List<TourismResponse>>> {
+        val resultData = PublishSubject.create<ApiResponse<List<TourismResponse>>>()
+
+        val client = apiService.getList()
+
+        client.subscribeOn(Schedulers.computation())
+            .observeOn(AndroidSchedulers.mainThread())
+            .take(1)
+            .subscribe({ res ->
+                val dataArray = res.places
+                resultData.onNext(if (dataArray.isNotEmpty()) ApiResponse.Success(dataArray) else ApiResponse.Empty)
+            }, { err ->
+                resultData.onNext(ApiResponse.Error(err.message.toString()))
+            })
+
+        return resultData.toFlowable(BackpressureStrategy.BUFFER)
+    }
+}
+
