+++ core/src/main/java/app/culturedev/core/data/source/remote/network/ApiService.kt
@@ -1,0 +1,10 @@
+package app.culturedev.core.data.source.remote.network
+
+import app.culturedev.core.data.source.remote.response.ListTourismResponse
+import io.reactivex.rxjava3.core.Flowable
+import retrofit2.http.GET
+
+interface ApiService {
+    @GET("list")
+    fun getList(): Flowable<ListTourismResponse>
+}
