+++ core/src/main/java/app/culturedev/core/data/source/remote/response/ListTourismResponse.kt
@@ -1,0 +1,14 @@
+package app.culturedev.core.data.source.remote.response
+
+import com.google.gson.annotations.SerializedName
+
+data class ListTourismResponse(
+    @field:SerializedName("error")
+    val error: Boolean,
+
+    @field:SerializedName("message")
+    val message: String,
+
+    @field:SerializedName("places")
+    val places: List<TourismResponse>
+)
