+++ core/src/main/java/app/culturedev/core/di/AppModule.kt
@@ -1,0 +1,17 @@
+package app.culturedev.core.di
+
+import app.culturedev.core.domain.interactor.TourismInteractor
+import app.culturedev.core.domain.usecase.TourismUseCase
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+import dagger.hilt.android.scopes.ViewModelScoped
+
+@Module
+@InstallIn(ViewModelComponent::class)
+abstract class AppModule {
+    @Binds
+    @ViewModelScoped
+    abstract fun provideTourismUseCase(tourismInteractor: TourismInteractor): TourismUseCase
+}
