+++ core/src/main/java/app/culturedev/core/di/RepositoryModule.kt
@@ -1,0 +1,14 @@
+package app.culturedev.core.di
+
+import app.culturedev.core.domain.repository.ITourismRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module(includes = [NetworkModule::class, DatabaseModule::class])
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+    @Binds
+    abstract fun provideRepository(tourismRepository: app.culturedev.core.data.TourismRepository): ITourismRepository
+}
