+++ core/src/main/java/app/culturedev/core/domain/model/Tourism.kt
@@ -1,0 +1,17 @@
+package app.culturedev.core.domain.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Tourism(
+    val tourismId: String,
+    val name: String,
+    val description: String,
+    val address: String,
+    val latitude: Double,
+    val longitude: Double,
+    val like: Int,
+    val image: String,
+    val isFavorite: Boolean
+) : Parcelable
