+++ core/src/main/java/app/culturedev/core/utils/DataMapper.kt
@@ -1,0 +1,53 @@
+package app.culturedev.core.utils
+
+import app.culturedev.core.data.source.local.entity.TourismEntity
+import app.culturedev.core.data.source.remote.response.TourismResponse
+import app.culturedev.core.domain.model.Tourism
+
+object DataMapper {
+    fun mapResponsesToEntities(input: List<TourismResponse>): List<TourismEntity> {
+        val tourismList = ArrayList<TourismEntity>()
+        input.map {
+            val tourism = TourismEntity(
+                tourismId = it.id,
+                description = it.description,
+                name = it.name,
+                address = it.address,
+                latitude = it.latitude,
+                longitude = it.longitude,
+                like = it.like,
+                image = it.image,
+                isFavorite = false
+            )
+            tourismList.add(tourism)
+        }
+        return tourismList
+    }
+
+    fun mapEntitiesToDomain(list: List<TourismEntity>): List<Tourism> =
+        list.map {
+            Tourism(
+                tourismId = it.tourismId,
+                description = it.description,
+                name = it.name,
+                address = it.address,
+                latitude = it.latitude,
+                longitude = it.longitude,
+                like = it.like,
+                image = it.image,
+                isFavorite = it.isFavorite
+            )
+        }
+
+    fun mapDomainToEntity(input: Tourism) = TourismEntity(
+        tourismId = input.tourismId,
+        description = input.description,
+        name = input.name,
+        address = input.address,
+        latitude = input.latitude,
+        longitude = input.longitude,
+        like = input.like,
+        image = input.image,
+        isFavorite = input.isFavorite
+    )
+}
